+++ components/ContactSection/interfaces.ts
@@ -1,11 +1,9 @@
-
-interface Validation {
-    type: "required" | "email" | "minLength" | "maxLength" | "min" | "max" | "pattern";
-    value?: string | number;
+type Validation = {
+    regex: string;
     message: string;
 }
 
-interface Field {
+export interface Field {
     id: string;
     name: string;
     label: string;
@@ -17,17 +15,3 @@ interface Field {
     disabled?: boolean;
     groupWithNext?: boolean;
 }
-
-interface TextareaFieldProps extends Field {
-    characterLimit?: number;
-}
-
-interface ButtonFieldProps extends Field {
-    type: "submit" | "reset" | "button";
-}
-
-interface InputFieldProps extends Field {
-    type: "text" | "email" | "password" | "number" | "tel" | "url";
-}
-
-export type { Field, TextareaFieldProps, ButtonFieldProps, InputFieldProps };
